fixed storybook, components now in app folder
