+++ app/components/competition-card.stories.tsx
