+++ app/components/competition-card.tsx
