+++ app/components/player-card.tsx
@@ -1,11 +1,10 @@
 import { MilitaryTech, TrendingUp, TrendingUpOutlined } from "@mui/icons-material";
-import { Box, ButtonBase, Grid, Typography } from "@mui/material";
+import { Avatar, ButtonBase, Grid, Typography } from "@mui/material";
 import React from "react";
-import MaetIcon from "../components/maet-icon";
+import MaetIcon from "./maet-icon";
 
-export type TeamCardProps = {
+export type PlayerCardProps = {
     name: string,
-    location: string,
     position: number,
     medalColor?: string,
     score: number,
@@ -16,7 +15,7 @@ export type ScoreChange = {
     color: string,
     magnitude: number
 }
-const TeamCard = (props: TeamCardProps) => {
+const PlayerCard = (props: PlayerCardProps) => {
 	return (
 		<Grid
 			container
@@ -47,11 +46,9 @@ const TeamCard = (props: TeamCardProps) => {
 			<Grid
 				item
 				container
-				xs={4}
+				xs={2}
 			>
-				<ButtonBase sx={{ width: 70, height: 70, border: 1, borderRadius: 1, 
-					borderColor: "#f5f5f4", backgroundColor: "purple", m: 1 }}>
-				</ButtonBase>
+				<Avatar sx={{mt: 1, mb: 1}}></Avatar>
 			</Grid>
 			<Grid
 				item
@@ -59,13 +56,12 @@ const TeamCard = (props: TeamCardProps) => {
 				xs={3}
 				direction="column"
 			>
-				<Typography>South Bend</Typography>
-				<Typography sx={{fontWeight: 700}}>Hotshots</Typography>
+				<Typography sx={{fontWeight: 700}}>Jay Boog</Typography>
 			</Grid>
 			<Grid
 				item
 				container
-				xs={3}
+				xs={5}
 				direction="column"
 				alignItems="flex-end"
 				justifyContent="flex-end"
@@ -77,10 +73,9 @@ const TeamCard = (props: TeamCardProps) => {
 					direction="row"
 					alignItems="flex-end"
 					justifyContent="flex-end"
-					display="flex"
 				>
 					<MaetIcon sx={{mr: 0.5}}></MaetIcon>
-					<Typography sx={{fontWeight: 700, mr: 0.5}}>906</Typography>
+					<Typography sx={{fontWeight: 700, mr: 0.5}}>1111</Typography>
 				</Grid>
 				<Grid
 					container
@@ -91,11 +86,11 @@ const TeamCard = (props: TeamCardProps) => {
 					justifyContent="flex-end"
 				>   
 					<TrendingUp sx={{color: "green",  fontSize: 20, mr: 0.5}}></TrendingUp>
-					<Typography sx={{color: "green", mr: 0.5}}>+218</Typography>
+					<Typography sx={{color: "green", mr: 0.5}}>+237</Typography>
 				</Grid>
 			</Grid>
 		</Grid>
 	);
 };
 
-export default TeamCard;
+export default PlayerCard;
