+++ app/components/team-card.tsx
@@ -1,10 +1,11 @@
 import { MilitaryTech, TrendingUp, TrendingUpOutlined } from "@mui/icons-material";
-import { Avatar, ButtonBase, Grid, Typography } from "@mui/material";
+import { Box, ButtonBase, Grid, Typography } from "@mui/material";
 import React from "react";
-import MaetIcon from "../components/maet-icon";
+import MaetIcon from "./maet-icon";
 
-export type PlayerCardProps = {
+export type TeamCardProps = {
     name: string,
+    location: string,
     position: number,
     medalColor?: string,
     score: number,
@@ -15,7 +16,7 @@ export type ScoreChange = {
     color: string,
     magnitude: number
 }
-const PlayerCard = (props: PlayerCardProps) => {
+const TeamCard = (props: TeamCardProps) => {
 	return (
 		<Grid
 			container
@@ -46,9 +47,11 @@ const PlayerCard = (props: PlayerCardProps) => {
 			<Grid
 				item
 				container
-				xs={2}
+				xs={4}
 			>
-				<Avatar sx={{mt: 1, mb: 1}}></Avatar>
+				<ButtonBase sx={{ width: 70, height: 70, border: 1, borderRadius: 1, 
+					borderColor: "#f5f5f4", backgroundColor: "purple", m: 1 }}>
+				</ButtonBase>
 			</Grid>
 			<Grid
 				item
@@ -56,12 +59,13 @@ const PlayerCard = (props: PlayerCardProps) => {
 				xs={3}
 				direction="column"
 			>
-				<Typography sx={{fontWeight: 700}}>Jay Boog</Typography>
+				<Typography>South Bend</Typography>
+				<Typography sx={{fontWeight: 700}}>Hotshots</Typography>
 			</Grid>
 			<Grid
 				item
 				container
-				xs={5}
+				xs={3}
 				direction="column"
 				alignItems="flex-end"
 				justifyContent="flex-end"
@@ -73,9 +77,10 @@ const PlayerCard = (props: PlayerCardProps) => {
 					direction="row"
 					alignItems="flex-end"
 					justifyContent="flex-end"
+					display="flex"
 				>
 					<MaetIcon sx={{mr: 0.5}}></MaetIcon>
-					<Typography sx={{fontWeight: 700, mr: 0.5}}>1111</Typography>
+					<Typography sx={{fontWeight: 700, mr: 0.5}}>906</Typography>
 				</Grid>
 				<Grid
 					container
@@ -86,11 +91,11 @@ const PlayerCard = (props: PlayerCardProps) => {
 					justifyContent="flex-end"
 				>   
 					<TrendingUp sx={{color: "green",  fontSize: 20, mr: 0.5}}></TrendingUp>
-					<Typography sx={{color: "green", mr: 0.5}}>+237</Typography>
+					<Typography sx={{color: "green", mr: 0.5}}>+218</Typography>
 				</Grid>
 			</Grid>
 		</Grid>
 	);
 };
 
-export default PlayerCard;
+export default TeamCard;
